refactor: . pull isFilenameUnique() up
